organized page, summary and history page misplacement fixed.
