+++ lib/features/salary/presentation/screens/salary_screen.dart
@@ -24,6 +24,16 @@ class _SalaryScreenState extends State<SalaryScreen> {
     return Scaffold(
       appBar: AppBar(
         title: const Text('Salary'),
+        actions: [
+          IconButton(
+            icon: const Icon(Icons.refresh),
+            onPressed: () async {
+              await context.read<SalaryController>().loadCurrentMonthSalary();
+              await context.read<SalaryController>().loadMonthlyHistory();
+            },
+            tooltip: 'Refresh data',
+          ),
+        ],
       ),
       body: Consumer<SalaryController>(
         builder: (context, controller, child) {
@@ -32,162 +42,196 @@ class _SalaryScreenState extends State<SalaryScreen> {
           }
 
           final currentSalary = controller.currentMonthSalary;
-          final selectedSalary = controller.selectedMonthSalary;
-          final isCurrentMonth = controller.selectedMonth.year == DateTime.now().year &&
-              controller.selectedMonth.month == DateTime.now().month;
+          if (currentSalary == null) {
+            return const Center(child: Text('No salary data available'));
+          }
 
-          return RefreshIndicator(
-            onRefresh: () async {
-              await controller.loadCurrentMonthSalary();
-              await controller.loadMonthlyHistory();
-            },
-            child: SingleChildScrollView(
-              physics: const AlwaysScrollableScrollPhysics(),
+          return ListView(
+            padding: const EdgeInsets.all(16.0),
+            children: [
+              // Today's Earnings Card
+              Card(
+                elevation: 4,
+                child: Padding(
               padding: const EdgeInsets.all(16.0),
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  // Month selector
-                  MonthSelector(
-                    selectedMonth: controller.selectedMonth,
-                    onMonthSelected: controller.selectMonth,
-                  ),
-                  const SizedBox(height: 24),
-
-                  // Hours progress gauge
-                  if (selectedSalary != null) ...[
-                    SizedBox(
-                      height: 200,
-                      child: HoursProgressGauge(
-                        actualMinutes: selectedSalary['actualMinutes'] as int,
-                        expectedMinutes: selectedSalary['expectedMinutes'] as int,
-                        label: 'Hours Progress',
-                      ),
-                    ),
-                    const SizedBox(height: 24),
-                  ],
-
-                  // Salary cards
-                  if (selectedSalary != null) ...[
-                    GridView.count(
-                      crossAxisCount: 2,
-                      shrinkWrap: true,
-                      physics: const NeverScrollableScrollPhysics(),
-                      mainAxisSpacing: 16,
-                      crossAxisSpacing: 16,
-                      childAspectRatio: 1.2,
+                      Row(
                       children: [
-                        SalaryCard(
-                          title: 'Total Earnings',
-                          value: controller.formatCurrency(
-                            selectedSalary['totalEarnings'] as double,
+                          Icon(
+                            Icons.today,
+                            color: Theme.of(context).colorScheme.primary,
+                            size: 28,
                           ),
-                          icon: Icons.attach_money,
-                          color: Colors.green,
+                          const SizedBox(width: 8),
+                          Text(
+                            'Today\'s Earnings',
+                            style: Theme.of(context).textTheme.titleLarge,
                         ),
-                        SalaryCard(
-                          title: 'After Insurance',
-                          value: controller.formatCurrency(
-                            selectedSalary['earningsAfterInsurance'] as double,
+                        ],
+                      ),
+                      const SizedBox(height: 16),
+                      Center(
+                        child: Text(
+                          controller.formatCurrency(currentSalary['todayEarnings'] as double),
+                          style: Theme.of(context).textTheme.headlineMedium?.copyWith(
+                            color: Theme.of(context).colorScheme.primary,
+                            fontWeight: FontWeight.bold,
                           ),
-                          icon: Icons.health_and_safety,
-                          color: Colors.blue,
                         ),
-                        SalaryCard(
-                          title: 'Overtime Pay',
-                          value: controller.formatCurrency(
-                            selectedSalary['overtimePay'] as double,
+                      ),
+                      const SizedBox(height: 8),
+                      Center(
+                        child: Text(
+                          'Based on ${controller.formatDuration(currentSalary['todayMinutes'] as int)} worked today',
+                          style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                            color: Colors.grey[600],
                           ),
-                          icon: Icons.timer,
-                          color: Colors.orange,
                         ),
-                        SalaryCard(
-                          title: 'Daily Rate',
-                          value: controller.formatCurrency(
-                            selectedSalary['dailyRate'] as double,
-                          ),
-                          icon: Icons.calendar_today,
-                          color: Colors.purple,
                         ),
                       ],
                     ),
-                    const SizedBox(height: 24),
-                  ],
-
-                  // Monthly history chart
-                  if (controller.monthlyHistory.isNotEmpty) ...[
-                    Text(
-                      'Monthly History',
-                      style: Theme.of(context).textTheme.titleLarge,
+                ),
                     ),
                     const SizedBox(height: 16),
-                    MonthlySalaryChart(
-                      monthlyData: controller.monthlyHistory,
-                    ),
-                    const SizedBox(height: 24),
-                  ],
 
-                  // Work days summary
-                  if (selectedSalary != null) ...[
+              // Current Month Card
                     Card(
+                elevation: 4,
                       child: Padding(
                         padding: const EdgeInsets.all(16.0),
                         child: Column(
                           crossAxisAlignment: CrossAxisAlignment.start,
                           children: [
+                      Row(
+                        children: [
+                          Icon(
+                            Icons.calendar_month,
+                            color: Theme.of(context).colorScheme.primary,
+                            size: 28,
+                          ),
+                          const SizedBox(width: 8),
                             Text(
-                              'Work Days Summary',
+                            'Current Month (${controller.getCurrentMonthName()})',
                               style: Theme.of(context).textTheme.titleLarge,
+                          ),
+                        ],
                             ),
                             const SizedBox(height: 16),
-                            _buildSummaryRow(
-                              'Expected Work Days',
-                              selectedSalary['expectedWorkDays'].toString(),
+                      _buildSalaryInfoRow(
+                        'Monthly Salary:',
+                        controller.formatCurrency(currentSalary['monthlySalary'] as double),
+                      ),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Expected Hours:',
+                        '${currentSalary['expectedHours']} hours (${currentSalary['workDaysCount']} days × ${currentSalary['dailyHours']}h)',
                             ),
-                            _buildSummaryRow(
-                              'Off Days',
-                              selectedSalary['offDaysCount'].toString(),
-                            ),
-                            _buildSummaryRow(
-                              'Hours Worked',
-                              controller.formatDuration(
-                                selectedSalary['actualMinutes'] as int,
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Off Days (Excused):',
+                        '${currentSalary['offDaysCount']} days',
+                        textColor: Colors.grey[600],
+                      ),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Weekends:',
+                        '${currentSalary['nonWorkingDaysCount']} days',
+                        textColor: Colors.grey[600],
+                      ),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Total Days Off:',
+                        '${currentSalary['totalDaysOff']} days',
+                        textColor: Theme.of(context).colorScheme.primary,
+                        primary: true,
+                      ),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Daily Rate:',
+                        controller.formatCurrency(currentSalary['dailyRate'] as double),
+                      ),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Hourly Rate:',
+                        '${controller.formatCurrency(currentSalary['hourlyRate'] as double)}/hour',
+                      ),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Overtime Pay:',
+                        controller.formatCurrency(currentSalary['overtimePay'] as double),
+                        textColor: (currentSalary['overtimePay'] as double) > 0 ? Colors.green : null,
+                      ),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Total Earnings:',
+                        controller.formatCurrency(currentSalary['totalEarnings'] as double),
+                        primary: true,
+                        textColor: Theme.of(context).colorScheme.primary,
                               ),
-                            ),
-                            if (selectedSalary['overtimeMinutes'] as int > 0)
-                              _buildSummaryRow(
-                                'Overtime Hours',
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'After Insurance (92%):',
+                        controller.formatCurrency(currentSalary['earningsAfterInsurance'] as double),
+                        textColor: Colors.grey[600],
+                      ),
+                      const SizedBox(height: 16),
+                      const Divider(),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Hours Worked:',
+                        controller.formatDuration(currentSalary['actualMinutes'] as int),
+                      ),
+                      const SizedBox(height: 8),
+                      _buildSalaryInfoRow(
+                        'Hours Left to Do:',
                                 controller.formatDuration(
-                                  selectedSalary['overtimeMinutes'] as int,
+                          (currentSalary['expectedMinutes'] as int) - (currentSalary['actualMinutes'] as int),
                                 ),
+                        textColor: ((currentSalary['expectedMinutes'] as int) - (currentSalary['actualMinutes'] as int)) < 0 
+                            ? Colors.green 
+                            : Colors.orange,
                               ),
                           ],
                         ),
                       ),
                     ),
                   ],
-                ],
-              ),
-            ),
           );
         },
       ),
     );
   }
 
-  Widget _buildSummaryRow(String label, String value) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 4.0),
-      child: Row(
+  Widget _buildSalaryInfoRow(
+    String label,
+    String value, {
+    bool primary = false,
+    Color? textColor,
+  }) {
+    return Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
-          Text(label),
+        Text(
+          label,
+          style: primary
+              ? Theme.of(context).textTheme.titleMedium
+              : Theme.of(context).textTheme.bodyMedium?.copyWith(
+                    fontWeight: FontWeight.bold,
+                  ),
+        ),
           Text(
             value,
-            style: const TextStyle(fontWeight: FontWeight.bold),
+          style: primary
+              ? Theme.of(context).textTheme.titleMedium?.copyWith(
+                    color: textColor ?? Theme.of(context).colorScheme.primary,
+                  )
+              : Theme.of(context).textTheme.bodyMedium?.copyWith(
+                    color: textColor,
+                  ),
           ),
         ],
-      ),
     );
   }
 }
